Conflicts:
	src/uke9oppgave2/Uke9Oppgave2.java

+++ src/insurancecompany/people/Employee.java
@@ -24,7 +24,7 @@ public abstract class Employee extends Person {
     
     private int employeeId;
     
-    public Employee(String firstname, String lastname, int personalNumber,
+    public Employee(String firstname, String lastname, long personalNumber,
             String email, Address address, int phone) {
         super(firstname, lastname, personalNumber, email, address, phone);
         employeeId = nextEmployeeId++;
